权限控制出错解决

+++ src/group/GroupManager.java
@@ -27,8 +27,8 @@ import database.DB;
 public class GroupManager {
 	protected static Log logger = LogFactory.getLog(GroupManager.class);
 	private static GroupManager instance ;
-	
-	private HashMap<Integer,List<Group>> map = new HashMap<Integer,List<Group>>();
+	 
+	//private HashMap<Integer,List<Group>> map = new HashMap<Integer,List<Group>>();
 	
 	private GroupManager(){
 	}
@@ -42,10 +42,10 @@ public class GroupManager {
 		return instance ;
 	}
 	
-	private void init(){
+	/*private void init(){
 		GroupManager.getInstance().map = new HashMap<Integer,List<Group>>() ;
 	}
-	  
+	   
 	public List<Group> getListGroupFromPemission(int type){
 		List<Group> list = map.get(type);
 		if(list == null ){
@@ -53,7 +53,7 @@ public class GroupManager {
 			map.put(type, list);
 		}
 		return list ;
-	}  
+	} */ 
 	
 
 	public static boolean getName(String c){
@@ -493,17 +493,21 @@ logger.info(sql);
 				return -2 ;
 			}
 			int count = 0 ;
-			
+			 
+			Group g = GroupManager.getGroup(Integer.valueOf(id));
 			
 			String sql = "delete from mdgroup where id = " + id;
 			
-			String sql1 = "delete from mdgroup where id  in  (select groupid from mdrelategroup  where pgroupid = "+id+") ";
+			if(g.getPtype() == 4){
+				String sql1 = "delete from mdgroup where id  in  (select groupid from mdrelategroup  where pgroupid = "+id+") ";
+				
+				String sql2 = "delete from mdbranch where bname =  (select groupname  from mdgroup where id = "+id+" )" ;
+				 
+		        sqls.add(sql2);  
+		        
+		       sqls.add(sql1); 
+			}
 			
-			String sql2 = "delete from mdbranch where bname =  (select groupname  from mdgroup where id = "+id+" )" ;
-			
-	        sqls.add(sql2); 
-	        
-	        sqls.add(sql1); 
 	            
             sqls.add(sql);
